refactor: form state 상수화
